+++ Assets/GameAssets/Unlockable.cs
@@ -1,5 +1,7 @@
+using System.Numerics;
+
 public interface Unlockable
 {
     public void Unlock();
-    public int GetUnlockCost();
+    public BigInteger GetUnlockCost();
 }
